Checked with James, the reason for older versions of Spring is because it harvested from OLS3. Thus, no reason not be on newer versions.

+++ backend/src/main/java/uk/ac/ebi/spot/ols/repository/solr/OlsSolrClient.java
@@ -74,7 +74,7 @@ public class OlsSolrClient {
     public QueryResponse runSolrQuery(SolrQuery query, Pageable pageable) {
 
 	if(pageable != null) {
-		query.setStart(pageable.getOffset());
+		query.setStart((int)pageable.getOffset());
 		query.setRows(pageable.getPageSize());
 	}
